pyproject.toml
[project]
name = "windmark"
version = "0.1.0"
description = "Multi-Structure Sequence Modeling"
readme = "docs/README.md"
requires-python = ">=3.11"
dependencies = [
    "flytekit==1.14.0b6",
    "torch==2.3.0", # torch>=2.4.0 has a GPU memory leak ?
    "lightning>=2.1.3",
    "tensordict>=0.3.0",
    "jaxtyping>=0.2.25",
    "beartype>=0.17.0",
    "torchmetrics>=1.3.2",
    "numpy>=1.26.4",
    "polars>=1.1.0",
    "pytdigest>=0.1.4",
    "pyarrow>=15.0.2",
    "pydantic>=2.6.4",
    "faker>=24.4.0",
    "hydra-core>=1.3.2",
    "mpire>=2.10.2",
    "msgspec>=0.18.6",
    "rich>=13.7.0",
    "textual>=0.72.0",
    "wandb>=0.18.3",
]

[tool.setuptools]
packages = ["windmark"]

[tool.uv]
dev-dependencies = [
    "interrogate>=1.7.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "nbformat>=5.10.4",
    "pre-commit>=4.0.0",
    "pydoctest>=0.2.1",
    "pyjust>=1.18.1",
    "pysort>=1.0.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.9",
    "semver>=3.0.2",
]

[tool.ruff]
lint.extend-ignore = ["F722"]
line-length = 120

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-property-decorators = true
ignore-module = true
ignore-setters = true
